Extending BackwardCompatibility
Now onto Migration plan and unit test of that
