+++ src/main/java/hudson/plugins/accurev/extensions/impl/ReferenceTreeCheckout.java
@@ -1,12 +1,19 @@
 package hudson.plugins.accurev.extensions.impl;
 
 import hudson.Extension;
+import hudson.model.Job;
+import hudson.model.TaskListener;
+import hudson.plugins.accurev.AccurevSCM;
 import hudson.plugins.accurev.extensions.AccurevSCMExtension;
 import hudson.plugins.accurev.extensions.AccurevSCMExtensionDescriptor;
+import jenkins.plugins.accurev.AccurevClient;
+import jenkins.plugins.accurev.AccurevException;
+import jenkins.plugins.accurev.UpdateCommand;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
 
 import javax.annotation.Nonnull;
+import java.io.IOException;
 
 public class ReferenceTreeCheckout extends AccurevSCMExtension {
 
@@ -34,6 +41,11 @@ public class ReferenceTreeCheckout extends AccurevSCMExtension {
     @DataBoundSetter
     public void setCleanReferenceTree(boolean cleanReferenceTree) {
         this.cleanReferenceTree = cleanReferenceTree;
+    }
+
+    @Override
+    public void decorateUpdateCommand(AccurevSCM scm, Job project, AccurevClient accurev, TaskListener listener, UpdateCommand cmd) throws IOException, InterruptedException, AccurevException {
+
     }
 
     @Extension
